Moved `ProducerJWKKey` and `ClientJWKKey` in `key.ts`
